improve permissions check logic
