+++ libs/domain/picking/offline/services/bapi-push-notification-default.service.spec.ts
@@ -10,6 +10,15 @@ class MockPushService implements Partial<PushService> {
   unsubscribe = vi.fn().mockImplementation(() => of(null));
 }
 
+const defaultPermissions = {
+  notification: 'granted',
+  'background-sync': 'granted',
+} as Record<string, string>;
+
+const mockPermissions = (permissions = defaultPermissions) => ({
+  query: ({ name }: { name: string }) => ({ state: permissions[name] }),
+});
+
 class MockBapiPushNotificationAdapter
   implements Partial<BapiPushNotificationAdapter>
 {
@@ -31,6 +40,12 @@ describe('BapiPushNotificationDefaultService', () => {
   let storage: MockStorageService;
 
   beforeEach(() => {
+    vi.stubGlobal('navigator', {
+      serviceWorker: {},
+      permissions: mockPermissions(),
+    });
+    vi.stubGlobal('window', { SyncManager: {} });
+
     const testInjector = createInjector({
       providers: [
         {
@@ -75,11 +90,6 @@ describe('BapiPushNotificationDefaultService', () => {
 
   describe('when initSubscription is called', () => {
     let subscription: Subscription;
-    const callback = vi.fn();
-
-    beforeEach(() => {
-      subscription = service.initSubscription().subscribe(callback);
-    });
 
     afterEach(() => {
       if (subscription) {
@@ -88,11 +98,12 @@ describe('BapiPushNotificationDefaultService', () => {
     });
 
     it('should subscribe and send subscription', () => {
-      expect(callback).toHaveBeenCalled();
-      expect(pushService.subscribe).toHaveBeenCalled();
-      expect(adapter.sendSubscription).toHaveBeenCalled();
-      expect(storage.get).toHaveBeenCalledWith(subscriptionFlagKey);
-      expect(storage.set).toHaveBeenCalledWith(subscriptionFlagKey, true);
+      service.initSubscription().subscribe(() => {
+        expect(pushService.subscribe).toHaveBeenCalled();
+        expect(adapter.sendSubscription).toHaveBeenCalled();
+        expect(storage.get).toHaveBeenCalledWith(subscriptionFlagKey);
+        expect(storage.set).toHaveBeenCalledWith(subscriptionFlagKey, true);
+      });
     });
 
     it('should handle error when subscription already exists', () => {
@@ -129,6 +140,86 @@ describe('BapiPushNotificationDefaultService', () => {
       expect(callback).toHaveBeenCalled();
       expect(pushService.unsubscribe).toHaveBeenCalled();
       expect(storage.remove).toHaveBeenCalledWith(subscriptionFlagKey);
+    });
+  });
+
+  describe('error handling', () => {
+    describe('when service-worker API is not supported', () => {
+      beforeEach(() => {
+        vi.stubGlobal('navigator', {});
+      });
+
+      it('should throw an error', () => {
+        service
+          .initSubscription()
+          .subscribe({
+            error: (e) =>
+              expect(e.message).toBe(
+                'Browser does not support service-worker API'
+              ),
+          })
+          .unsubscribe();
+      });
+    });
+
+    describe('when syncManager API is not supported', () => {
+      beforeEach(() => {
+        vi.stubGlobal('window', {});
+      });
+
+      it('should throw an error', () => {
+        service
+          .initSubscription()
+          .subscribe({
+            error: (e) =>
+              expect(e.message).toBe(
+                'Browser does not support background sync API'
+              ),
+          })
+          .unsubscribe();
+      });
+    });
+
+    describe('when notifications are not allowed', () => {
+      beforeEach(() => {
+        vi.stubGlobal('navigator', {
+          serviceWorker: {},
+          permissions: mockPermissions({ notification: 'denied' }),
+        });
+      });
+
+      it('should throw an error', () => {
+        service
+          .initSubscription()
+          .subscribe({
+            error: (e) =>
+              expect(e.message).toBe(
+                'Permission to accept push notifications is not granted. Check the browser configuration or reset the permission'
+              ),
+          })
+          .unsubscribe();
+      });
+    });
+
+    describe('when background sync is not allowed', () => {
+      beforeEach(() => {
+        vi.stubGlobal('navigator', {
+          serviceWorker: {},
+          permissions: mockPermissions({ 'background-sync': 'denied' }),
+        });
+      });
+
+      it('should throw an error', () => {
+        service
+          .initSubscription()
+          .subscribe({
+            error: (e) =>
+              expect(e.message).toBe(
+                'Permission to perform background sync is not granted. Check the browser configuration or reset the permission'
+              ),
+          })
+          .unsubscribe();
+      });
     });
   });
 });
